feat(socket-udp): publish the six tvOS/watchOS targets
:socket-udp published five Apple targets; a consumer on tvOS or watchOS had no
UdpSocket to bind. Now all eleven, matching root :socket and :network-monitor.

The build file blamed buffer-flow's Apple klib coverage. That was true at buffer
6.11.0 and is stale: buffer / buffer-flow / buffer-codec all publish klibs for
tvos{Arm64,SimulatorArm64,X64} and watchos{Arm64,SimulatorArm64,X64} at the pinned
6.25.0, and a metadata dump of buffer-flow watchosArm64 vs macosArm64 is identical.
Network.framework is watchOS 6+ / tvOS 12+ (K/N floors are 8.0 / 15.0), and root
:socket already ships NWConnection TCP on these same targets.

Two real blockers, both arm64_32 (watchosArm64 has 32-bit pointers, so size_t /
ssize_t / NSUInteger are 32-bit there and 64-bit everywhere else):

1. appleMain forces a compileAppleMainKotlinMetadata compilation that type-checks one
   source against every Apple target at once. The datagram channels' memcpy lengths,
   sizeOf<…>().convert() and content.length then become "numbers with different bit
   widths in at least two actual platforms" and metadata compilation fails — while each
   individual target compiles fine. Fixed the way root :socket and :network-monitor
   already do it: src/appleNativeImpl/kotlin added per-target via srcDir, no appleMain
   metadata compilation. A slim appleMain survives for appleSockAddrLayout and
   UdpConnectException, which are width-free and which :socket-quic-quiche's own
   appleMain resolves from metadata.
2. HostLoopback.apple.kt compared raw ssize_t returns against Long. Normalized both
   sides. Verified by reverting it: watchosArm64 fails with "Operator '!=' cannot be
   applied to 'Int' and 'Long'" while macosArm64 compiles the identical source clean.

Because every simulator lane is 64-bit, neither blocker was catchable by CI. Added a
compile-only build-apple step for the device targets (watchosArm64, tvosArm64,
iosArm64) so an ILP32 regression cannot land silently again.

Verification: all 11 targets compile and publishToMavenLocal emits klibs for all six
new ones; tvosSimulatorArm64Test and watchosSimulatorArm64Test each run 60 tests,
0 failures, 0 skipped — the same count as macosArm64.

Unblocks DitchOoM/webrtc#127. Closes #280.

+++ socket-udp/src/appleTest/kotlin/com/ditchoom/socket/udp/HostLoopback.apple.kt
@@ -76,7 +76,10 @@ private fun rawLoopbackCarries(size: Int): Boolean =
             val payload = allocArray<ByteVar>(size)
             for (i in 0 until size) payload[i] = 0x41
             val sent = sendto(sender, payload, size.convert(), 0, target.reinterpret(), addrLength.convert())
-            if (sent != size.toLong()) return false // the host refused it outright (EMSGSIZE)
+            // .toLong() on BOTH sides: sendto returns ssize_t, which K/N maps to Long on every Apple
+            // target except watchosArm64 (arm64_32), where it is Int. Comparing the raw return against
+            // a Long compiles everywhere else and fails only there.
+            if (sent.toLong() != size.toLong()) return false // the host refused it outright (EMSGSIZE)
 
             // Accepted — now find out whether it actually arrives. Silence here is the WSL2 shape: a
             // host that drops rather than refuses, and the case this whole seam exists to recognize.
@@ -87,7 +90,7 @@ private fun rawLoopbackCarries(size: Int): Boolean =
             if (poll(fds.ptr, 1.convert(), RECEIVE_TIMEOUT_MILLIS) <= 0) return false
 
             val landing = allocArray<ByteVar>(size + 1)
-            recvfrom(receiver, landing, (size + 1).convert(), 0, null, null) == size.toLong()
+            recvfrom(receiver, landing, (size + 1).convert(), 0, null, null).toLong() == size.toLong()
         } finally {
             close(sender)
             close(receiver)
